+++ src/recsys/data/EntityInteraction.h
@@ -46,13 +46,13 @@ public:
 	/// index EntityInteraction by entity ids
 	typedef vector<entity_interact_ptr> entity_interact_vec;
 	typedef shared_ptr<entity_interact_vec> entity_interact_vec_ptr;
-	typedef map<ushort,entity_interact_vec_ptr> type_interact_map;
-	typedef map<Entity::mapped_id_type,map<ushort,entity_interact_vec_ptr> > entity_type_interact_map;
+	typedef map<int8_t,entity_interact_vec_ptr> type_interact_map;
+	typedef map<Entity::mapped_id_type,map<int8_t,entity_interact_vec_ptr> > entity_type_interact_map;
 	typedef map<Entity::mapped_id_type,set<Entity::mapped_id_type> > id_id_map;
 public:
 	Entity::entity_ptr m_from_entity;
 	Entity::entity_ptr m_to_entity;
-	ushort m_type;
+	int8_t m_type;
 	/// direct value
 	float m_val;
 	js::Object m_desc;
@@ -62,11 +62,11 @@ public:
 	static entity_type_interact_map m_entity_type_interact_map;
 	static id_id_map m_id_id_map;
 protected:
-	Entity::entity_ptr _index_entity(string const& id, ushort type, js::Object const& val);
-	entity_interact_vec_ptr _create_vec_if_not_exist(ushort const& entType, size_t const& entId);
+	Entity::entity_ptr _index_entity(string const& id, int8_t type, js::Object const& val);
+	entity_interact_vec_ptr _create_vec_if_not_exist(int8_t const& entType, size_t const& entId);
 public:
-	EntityInteraction(ushort const& type, float val, js::Object const& desc = js::Object(), bool memoryMode = true);
-	EntityInteraction(ushort const& type,js::Object const& desc = js::Object(), bool memoryMode = true);
+	EntityInteraction(int8_t const& type, float val, js::Object const& desc = js::Object(), bool memoryMode = true);
+	EntityInteraction(int8_t const& type,js::Object const& desc = js::Object(), bool memoryMode = true);
 	inline void set_from_entity(Entity::entity_ptr const& fromEntityPtr){
 		m_from_entity = fromEntityPtr;
 	}
@@ -74,9 +74,9 @@ public:
 		m_to_entity = toEntityPtr;
 	}
 
-	void add_from_entity(string const& id, ushort type, js::Object const& val = js::Object());
-	void add_to_entity(string const& id, ushort type, js::Object const& val = js::Object());
-	inline ushort get_type() const{
+	void add_from_entity(string const& id, int8_t type, js::Object const& val = js::Object());
+	void add_to_entity(string const& id, int8_t type, js::Object const& val = js::Object());
+	inline int8_t get_type() const{
 		return m_type;
 	}
 	entity_interact_ptr index_if_not_exist();
@@ -84,8 +84,8 @@ public:
 public:
 	static SharedData init_shared_data();;
 	static bool entity_interact_exist(Entity::mapped_id_type const& fromId, Entity::mapped_id_type& toId, bool memoryMode);
-	static entity_interact_vec_ptr query(string const& entityName, ushort const& entityType, ushort const& intType, bool memoryMode = true, bool isFrom = true);
-	static entity_interact_vec_ptr query(size_t const& entityId, ushort const& intType, bool memoryMode = true, bool isFrom = true);
+	static entity_interact_vec_ptr query(string const& entityName, int8_t const& entityType, int8_t const& intType, bool memoryMode = true, bool isFrom = true);
+	static entity_interact_vec_ptr query(size_t const& entityId, int8_t const& intType, bool memoryMode = true, bool isFrom = true);
 };
 
 ostream& operator<<(ostream& oss, EntityInteraction const& rhs);
